Aggiornamento ulteriore per delle imprecisioni
Abbiamo sistemato il Codice di "public void Occorrenze()" nella classe "StatisticheDati". Inoltre abbiamo cambiato il tipo della variabile "TotaleAccessi" da int a double perché altrimenti quando andavamo a calcolarci le statistiche in alcuni casi ci restituiva null o solo zeri.

+++ ProgettoOOP/src/main/java/data/Dataset.java
@@ -255,7 +255,7 @@ public class Dataset implements Indice	//La Classe Dataset implementa l'Interfac
 		OsservazioneBreveIntensiva elemento=new OsservazioneBreveIntensiva(Dati[NOME_STRUTTURA],Dati[COMUNE]);
 		   try
 		   {
-			   elemento.setTotaleAccessi(Integer.parseInt(Dati[TOTALE_ACCESSI]));
+			   elemento.setTotaleAccessi(Double.parseDouble(Dati[TOTALE_ACCESSI]));
 		   }
 		   catch(NumberFormatException e)	//Se la conversione della stringa in numero fallisce viene lanciata un'eccezione 
 		   {
